+++ neocarta/connectors/bigquery/logs/models.py
@@ -12,3 +12,4 @@ class LogsExtractorCache(TypedDict):
     table_info: pd.DataFrame | None
     column_info: pd.DataFrame | None
     column_references_info: pd.DataFrame | None
+    cte_info: pd.DataFrame | None