+++ neocarta/ingest/rdbms/constraints.py
@@ -54,6 +54,13 @@ FOR (q:Query) REQUIRE q.id IS UNIQUE;
 
 query_id_key_constraint = query_id_unique_constraint.replace("UNIQUE", "NODE KEY")
 
+cte_id_unique_constraint = """
+CREATE CONSTRAINT cte_id_constraint IF NOT EXISTS
+FOR (c:CTE) REQUIRE c.id IS UNIQUE;
+"""
+
+cte_id_key_constraint = cte_id_unique_constraint.replace("UNIQUE", "NODE KEY")
+
 UNIQUE_CONSTRAINTS_LOOKUP = {
     NodeLabel.DATABASE: database_id_unique_constraint,
     NodeLabel.SCHEMA: schema_id_unique_constraint,
@@ -64,6 +71,7 @@ UNIQUE_CONSTRAINTS_LOOKUP = {
     NodeLabel.CATEGORY: category_id_unique_constraint,
     NodeLabel.BUSINESS_TERM: business_term_id_unique_constraint,
     NodeLabel.QUERY: query_id_unique_constraint,
+    NodeLabel.CTE: cte_id_unique_constraint,
 }
 
 KEY_CONSTRAINTS_LOOKUP = {
@@ -76,4 +84,5 @@ KEY_CONSTRAINTS_LOOKUP = {
     NodeLabel.CATEGORY: category_id_key_constraint,
     NodeLabel.BUSINESS_TERM: business_term_id_key_constraint,
     NodeLabel.QUERY: query_id_key_constraint,
+    NodeLabel.CTE: cte_id_key_constraint,
 }